fix: address review feedback -- CLI ScaleFactor handling, EXIF orientation, tests
- CLI: resolve_output_dimensions() is ScaleFactor-aware. The parser default
  "auto" arrives as ScaleFactor(1); the previous str() comparison never
  matched, so auto dimensions were wrongly sent through DimensionResolver.
  Both-auto now maps to (None, None) so the edit variant derives its ~1MP
  target from the last condition image; explicit scales resolve against that
  image; plain integers pass through.

- Edit variant: load condition images via open_oriented() so a portrait image
  stored landscape under EXIF Orientation 6 reaches both the vision
  preprocessor and the VAE the way it displays. RGBA is preserved for the VAE.

- Text encoder: cast scattered vision/deepstack embeds back to the stream
  dtype so mixed-precision checkpoints do not upcast the hidden states.

- Tests: add test_qwen_image_21_edit_exif.py (stubbed encoders, no weights)
  covering orientation propagation to both encoders plus RGBA alpha
  preservation, and a CLI dimension-resolution test covering
  auto/scale/integer flag combinations.

+++ src/mflux/models/qwen21/model/qwen21_text_encoder/qwen21_text_encoder.py
@@ -168,7 +168,7 @@ class Qwen21TextEncoder(nn.Module):
         # Scatter the vision embeds into the embedding sequence in order.
         mask_flat = image_mask[0]
         gather_index = mx.cumsum(mask_flat.astype(mx.int32)) - 1  # k-th True -> embed k
-        gathered = image_embeds[gather_index]
+        gathered = image_embeds[gather_index].astype(hidden_states.dtype)
         keep = mask_flat[:, None]  # (seq_len, 1)
         hidden_states = mx.where(keep[None, :, :], gathered[None, :, :], hidden_states)
         positions = Qwen21TextEncoder.build_mrope_positions(input_ids, image_mask, image_grid_thw)
@@ -187,7 +187,7 @@ class Qwen21TextEncoder(nn.Module):
             if layer_index < len(deepstack_embeds):
                 # inject at image positions only: expand the (n_image, hidden) deepstack
                 # features to full sequence length via the same gather index
-                ds_full = deepstack_embeds[layer_index][gather_index]
+                ds_full = deepstack_embeds[layer_index][gather_index].astype(hidden_states.dtype)
                 hidden_states = mx.where(keep[None, :, :], (hidden_states[0] + ds_full)[None, :, :], hidden_states)
 
         return hidden_states, image_mask
